api cleansing, moving handlers to the appropriaten module
